requirements.txt
flax>=0.11.1
jaxtyping>=0.3.2
numpy>=2.2.6
cyclopts>=3.22.5
attrs>=25.3.0
jax[cuda12]>=0.7.0
rope>=0.0.0
optax>=0.2.5
ipdb>=0.13.13
gymnasium>=1.2.0
tqdm>=4.67.1
pyyaml>=6.0.2
wandb>=0.21.1
ogbench>=1.1.5
opencv-python>=4.12.0.88
setuptools>=78.1.1
matplotlib>=3.10.5
scipy>=1.16.1
distrax>=0.1.7